Issue #347 - Initial commit. Add postprocessor to WithSameValue
This is an initial commit to add a postprocessor to WithSameValue to be
able to modify the returned Matcher. The approach based on a
postprocessor is required because of the way the super class matcher is
defined (once at the start and can't be modify latter thru the generated
interface).